First patch for foundry 0.8.6
