+++ package.json
@@ -17,6 +17,7 @@
   "devDependencies": {
     "@league-of-foundry-developers/foundry-vtt-types": "^0.7.10-0",
     "@types/async": "^3.2.6",
+    "@types/node": "^15.12.2",
     "@typescript-eslint/eslint-plugin": "^4.26.1",
     "@typescript-eslint/parser": "^4.26.1",
     "archiver": "^5.2.0",
