+++ src/module.json
@@ -4,8 +4,35 @@
 	"description": "Blood 'n Guts creates blood spray and splatter when tokens recieve damage, drawing to the floor and tokens, and drawing trails of blood after tokens who have taken damage. Many options are configurable.",
 	"author": "edzillion",
 	"version": "0.8.7",
-	"minimumCoreVersion": "0.6.0",
-	"compatibleCoreVersion": "0.7.9",
+	"minimumCoreVersion": "0.7.10",
+	"compatibleCoreVersion": "0.8.6",
+	"type": "module",
+	"socket": true,
+	"includes": [
+	  "./assets/**",
+	  "./lang/**",
+	  "./scripts/**",
+	  "./styles/**",
+	  "./templates/**",
+	  "./module.json",
+	  "./README.md",
+	  "./icons/**",
+	  "./packs/**"
+	],
+	"media": [
+	  {
+		"type": "icon",
+		"location": ""
+	  },
+	  {
+		"type": "cover",
+		"location": ""
+	  },
+	  {
+		"type": "screenshot",
+		"location": ""
+	  }
+	],	
 	"esmodules": [
 		"./blood-n-guts.js",
 		"./module/controls.js"
@@ -21,10 +48,17 @@
 			"path": "lang/en.json"
 		}
 	],
+	"dependencies": [
+		{
+			"name": "lib-wrapper"
+		}
+	],
 	"url": "https://github.com/edzillion/blood-n-guts",
 	"manifest": "https://raw.githubusercontent.com/edzillion/blood-n-guts/master/src/module.json",
 	"download": "https://github.com/edzillion/blood-n-guts/releases/download/v0.8.7/blood-n-guts-v0.8.7.zip",
 	"changelog": "https://github.com/edzillion/blood-n-guts/blob/v0.8.7/changelog.md",
 	"bugs": "https://github.com/edzillion/blood-n-guts/issues",
-	"allowBugReporter": true
-}
+	"allowBugReporter": true,
+	"manifestPlusVersion": "1.2.0"
+  }
+
