+++ tsconfig.json
@@ -1,27 +1,32 @@
 {
-	"compilerOptions": {
-    "moduleResolution": "node",
-		"noEmit": false,
-		"strict": false,
-		"target": "es2020",
-		"lib": [
-			"DOM",
-			"ES6",
-			"ES2020"
-    	],
-		"types": [
-			"node",
-      "async",
-      "@league-of-foundry-developers/foundry-vtt-types",
-		],
-    "resolveJsonModule" : true,
-    "baseUrl": "./",
-    "paths" : {
-      "src/*": ["./src/*"]
-    },
-    "sourceMap": true,
-    // "inlineSourceMap": true,
-    // "inlineSources": true
+	"compilerOptions":{
+	   "moduleResolution":"node",
+	   "noEmit":false,
+	   "strict":false,
+	   "target":"es2020",
+	   "lib":[
+		  "DOM",
+		  "ES6",
+		  "ES2020"
+	   ],
+	   "types":[
+		  "node",
+		  "async",
+		  "@league-of-foundry-developers/foundry-vtt-types"
+	   ],
+	   "resolveJsonModule":true,
+	   "baseUrl":"./",
+	   "paths":{
+		  "src/*":[
+			 "./src/*"
+		  ]
+	   },
+	   "sourceMap":true,
+	   "inlineSourceMap":false,
+	   "inlineSources":false
 	},
-  "exclude": ["node_modules", "dist"]
-}
+	"exclude":[
+	   "node_modules",
+	   "dist"
+	]
+ }
